for home. nothing interesting

+++ src/main/java/com/netcracker/solutions/kpi/persistence/dao/ApplicationFormDao.java
@@ -45,7 +45,7 @@ public interface ApplicationFormDao {
 
     int changeCurrentsAppFormStatus(Long fromIdStatus,Long toIdStatus);
 
-    Long getCountRecruitmentStudents(Short id);
+    Long getCountRecruitmentStudents(Long id);
 
     Long getApprovedStudentsByRecruitmentId(Short id);
