+++ src/main/java/com/netcracker/solutions/kpi/persistence/dto/scheduling/InterviewParametersDTO.java
@@ -4,7 +4,7 @@ package com.netcracker.solutions.kpi.persistence.dto.scheduling;
  * Created by olil0716 on 8/9/2016.
  */
 public class InterviewParametersDTO {
-    private Short recruitmentId;
+    private Long recruitmentId;
     private Short timeInterviewTech;
     private Short timeInterviewSoft;
 
@@ -12,16 +12,16 @@ public class InterviewParametersDTO {
     public InterviewParametersDTO() {
     }
 
-    public InterviewParametersDTO(Short recruitmentId, Short timeInterviewTech, Short timeInterviewSoft) {
+    public InterviewParametersDTO(Long recruitmentId, Short timeInterviewTech, Short timeInterviewSoft) {
         this.recruitmentId = recruitmentId;
         this.timeInterviewTech = timeInterviewTech;
         this.timeInterviewSoft = timeInterviewSoft;
     }
 
-    public Short getRecruitmentId() {
+    public Long getRecruitmentId() {
         return recruitmentId;
     }
-    public void setRecruitmentId(Short recruitmentId) {
+    public void setRecruitmentId(Long recruitmentId) {
         this.recruitmentId = recruitmentId;
     }
